+++ CHANGELOG.md
@@ -12,9 +12,13 @@
 
 ### Fixed
 
-- Kotlin JVM test dependencies applied in common source sets
-
 ### Updated
+
+## [0.5.0-beta.4] - 2023-07-22
+
+### Fixed
+
+- Kotlin JVM test dependencies applied in common source sets
 
 ## [0.5.0-beta.3] - 2023-07-21
 
@@ -489,7 +493,9 @@
 - `org.jetbrains.intellij.plugins:gradle-intellij-plugin -> 1.13.2`
 - `app.cash.molecule:molecule-gradle-plugin -> 0.8.0`
 
-[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-beta.3...HEAD
+[Unreleased]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-beta.4...HEAD
+
+[0.5.0-beta.4]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-beta.3...0.5.0-beta.4
 
 [0.5.0-beta.3]: https://github.com/JavierSegoviaCordoba/hubdle/compare/0.5.0-beta.2...0.5.0-beta.3